Added solution to P9. Palindrome Integers
Added solution to P9. Palindrome Integers

+++ P8.FactorialDivision/Program.cs
@@ -4,9 +4,9 @@ namespace P8.FactorialDivision
 {
     internal class Program
     {
-        static int CalculateFactorialOfANumber(int number)
+        static double CalculateFactorialOfANumber(double number)
         {
-            int fact = 1;
+            double fact = 1;
             for (int i = 1; i <= number; i++)
             {
                fact = fact * i;
@@ -14,20 +14,20 @@ namespace P8.FactorialDivision
             return fact;
         }
 
-        static void DivideTheTwoFactorials(int number, int number2)
+        static void DivideTheTwoFactorials(double number, double number2)
         {
-            int result = number / number2;
+            double result = number / number2;
             Console.WriteLine($"{result:f2}");
         }
    
 
         static void Main(string[] args)
         {
-            int number1 = int.Parse(Console.ReadLine());
-            int number2 = int.Parse(Console.ReadLine());
+            double number1 = double.Parse(Console.ReadLine());
+            double number2 = double.Parse(Console.ReadLine());
 
-            int factorial1 = CalculateFactorialOfANumber(number1);
-            int factorial2 = CalculateFactorialOfANumber(number2);
+            double factorial1 = CalculateFactorialOfANumber(number1);
+            double factorial2 = CalculateFactorialOfANumber(number2);
 
             DivideTheTwoFactorials(factorial1,factorial2);
